+++ plugin/src/main/java/io/github/zrdzn/minecraft/greatlifesteal/heart/HeartItem.java
@@ -5,7 +5,7 @@ import org.bukkit.inventory.ItemStack;
 
 public class HeartItem {
 
-    public int healthAmount;
+    public double healthAmount;
     public ItemStack result;
     public Map<Integer, ItemStack> ingredients;
